% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ochre_pal}
\alias{ochre_pal}
\title{ochre palette (discrete)}
\usage{
ochre_pal(n, palette = "namatjira_qual", alpha = 1)
}
\arguments{
\item{n}{number of colours when using the ramp}

\item{palette}{}

\item{alpha}{transpreancy}
}
\description{
ochre palette (discrete)
}
\examples{
library(scales)
show_col(ochre_pal()(10))

filled.contour(volcano,color.palette = ochre_pal(256), asp=1)

}
