% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_ochre}
\alias{scale_fill_ochre}
\title{#' Setup fill palette for ggplot2}
\usage{
scale_fill_ochre(..., palette = "namatjira_qual", discrete = TRUE)
}
\arguments{
\item{palette}{Choose from 'ochre_palettes' list}
}
\description{
#' Setup fill palette for ggplot2
}
