% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{viz_palette}
\alias{viz_palette}
\title{Show an ochre palette}
\usage{
viz_palette(pal)
}
\arguments{
\item{pal}{character, vector of (hexadecimal) colors characterizing a palette}
}
\description{
Show an ochre palette
}
