% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ochre_pal}
\alias{ochre_pal}
\title{ochre palette with ramped colours}
\usage{
ochre_pal(palette = "namatjira_qual", alpha = 1)
}
\arguments{
\item{palette}{Choose from 'ochre_palettes' list}

\item{alpha}{transparency}
}
\description{
ochre palette with ramped colours
}
\examples{
library(scales)
show_col(ochre_pal()(10))

filled.contour(volcano,color.palette = ochre_pal(), asp=1)

}
