% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{ochre_palettes}
\alias{ochre_palettes}
\title{Australian Colour Palettes}
\format{An object of class \code{list} of length 14.}
\usage{
ochre_palettes
}
\description{
A collection of colour palettes inspired by the landscape and wildlife of Australia
}
\examples{

# Make an x-y plot using the Namatjira palette
library(tidyverse)
df <- data.frame(x = rnorm(100, 0, 20), 
          y = rnorm(100, 0, 20), 
          cl = sample(letters[1:8], 100, replace=TRUE))
ggplot(df, aes(x, y, colour=cl, shape=cl)) + 
  geom_point(size=4) + scale_colour_ochre() +
  theme_bw() + theme(aspect.ratio=1) 

# Make a histogram using the McCrea Collins Street palette
ggplot(df, aes(x, fill=cl)) + geom_histogram() +
  scale_fill_ochre(palette="mccrea")

}
\keyword{datasets}
