% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{ochre_palettes}
\alias{ochre_palettes}
\title{Australia Color Palettes}
\format{An object of class \code{list} of length 12.}
\usage{
ochre_palettes
}
\description{
Australia Color Palettes
}
\keyword{datasets}
