% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_ochre}
\alias{scale_fill_ochre}
\title{#' Setup fill palette for ggplot2}
\usage{
scale_fill_ochre(..., palette = "namatjira_qual", discrete = TRUE,
  alpha = 1, reverse = TRUE)
}
\arguments{
\item{...}{parameters to \code{discrete_scale} or \code{scale_fill_gradientn}}

\item{palette}{Choose from 'ochre_palettes' list}

\item{discrete}{generate a discrete palette? (default: \code{FALSE} - generate continuous palette)}

\item{alpha}{pass through parameter to \code{viridis}}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
#' Setup fill palette for ggplot2
}
