% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{viz_palette}
\alias{viz_palette}
\title{Show an ochre palette}
\usage{
viz_palette(obj, nam, num = length(obj))
}
\arguments{
\item{obj}{character, vector of (hexadecimal) colors characterizing a palette}

\item{nam}{character, the name of the palette}

\item{num}{numeric, the number of colors to use}
}
\description{
Show an ochre palette
}
