% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{viz_palette}
\alias{viz_palette}
\title{Show a single palette}
\usage{
<<<<<<< HEAD
viz_palette(pal, ttl = "", num = length(pal))
=======
viz_palette(obj, name = NULL, num = length(obj))
>>>>>>> c437cc2260556a426203fea40e2aa9bc8ad63cd3
}
\arguments{
\item{pal}{character, vector of (hexadecimal) colours representing a palette}

<<<<<<< HEAD
\item{ttl}{character, title to be displayed (the name of the palette)}
=======
\item{name}{character, the name of the palette}
>>>>>>> c437cc2260556a426203fea40e2aa9bc8ad63cd3

\item{num}{numeric, the number of colours to display}
}
\description{
Display a single palette to see whether it meets your needs.
If no \code{num} parameter is given,
all the colours in the palette will be displayed.
If \code{num} is less than the number of colours in the palette,
then only the first \code{num} colours will be displayed.
If \code{num} is greater than the number of colours in the palette,
then that many colours will be generated by linear interpolation
over the vector of colours in the chosen palette.
}
\examples{
viz_palette(ochre_palettes$namatjira_qual)
viz_palette(ochre_palettes$namatjira_qual, "namatjira_qual")
viz_palette(ochre_palettes$namatjira_qual, "namatjira_qual first 4", 4)
viz_palette(ochre_palettes$namatjira_qual, "namatjira_qual interpolated to 25", 25)

}
